+++ frontend/apps/cruzcal/app/drop-down/drop-down.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import DropDown from './drop-down';
+
+describe('DropDown', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<DropDown name="MyDropDown" items={[]} />);
+    expect(baseElement).toBeTruthy();
+  });
+});
