+++ frontend/apps/cruzcal/app/drop-down/drop-down.tsx
@@ -3,11 +3,12 @@ import { FC, Fragment, useEffect, useRef, useState } from 'react'
 import { ChevronDownIcon } from '@heroicons/react/solid'
 
 /* eslint-disable-next-line */
-export interface SubjectDropDownProps {
-  subjects: string[]
-}
+export interface DropDownProps {
+  name: string;
+  items: string[];
+};
 
-export const SubjectDropDown: FC<SubjectDropDownProps> = (props) => {
+export const DropDown: FC<DropDownProps> = (props) => {
   const createMenuItem = (item) => {
     return (
       <Menu.Item>
@@ -24,14 +25,14 @@ export const SubjectDropDown: FC<SubjectDropDownProps> = (props) => {
     )
   };
 
-  const menuItems = props.subjects.map(createMenuItem);
+  const menuItems = props.items.map(createMenuItem);
 
   return (
     <div className="w-56 text-right fixed top-16">
       <Menu as="div" className="relative inline-block text-left">
         <div>
           <Menu.Button className="inline-flex justify-center w-full px-4 py-2 text-sm font-medium text-black bg-black rounded-md bg-opacity-20 hover:bg-opacity-30 focus:outline-none focus-visible:ring-2 focus-visible:ring-white focus-visible:ring-opacity-75">
-            Subject
+            {props.name}
             <ChevronDownIcon
               className="w-5 h-5 ml-2 -mr-1 text-violet-200 hover:text-violet-100"
               aria-hidden="true"
@@ -59,4 +60,4 @@ export const SubjectDropDown: FC<SubjectDropDownProps> = (props) => {
   );
 };
 
-export default SubjectDropDown;
+export default DropDown;
