--- frontend/apps/cruzcal/app/subject-drop-down/subject-drop-down.spec.tsx
@@ -1,10 +1,0 @@
-import { render } from '@testing-library/react';
-
-import SubjectDropDown from './subject-drop-down';
-
-describe('SubjectDropDown', () => {
-  it('should render successfully', () => {
-    const { baseElement } = render(<SubjectDropDown />);
-    expect(baseElement).toBeTruthy();
-  });
-});
