+++ frontend/nx.json
@@ -39,6 +39,10 @@
     "@nrwl/react": {
       "application": {
         "babel": true
+      },
+      "component": {
+        "style": "css",
+        "export": true
       }
     },
     "@nrwl/next": {
